feat: TagView 불필요한 기능 삭제 및 파일 통합
