+++ FilmGo/FilmGo/Presentation/Detail/TagsScrollView.swift
@@ -8,6 +8,8 @@
 import UIKit
 import SnapKit
 
+// MARK: - TagsScrollView
+
 final class TagsScrollView: UIScrollView {
     private let stackView: UIStackView = {
         let stackView = UIStackView()
@@ -32,7 +34,7 @@ final class TagsScrollView: UIScrollView {
 
     func updateTags(_ texts: [String], _ isDeletable: Bool = false) {
         texts.forEach {
-            let tagView = TagView(isDeletable: isDeletable)
+            let tagView = TagView()
             tagView.updateLabel($0)
             stackView.addArrangedSubview(tagView)
         }
@@ -44,7 +46,6 @@ private extension TagsScrollView {
         setAttributes()
         setHierarchy()
         setConstraints()
-        setDelegate()
     }
 
     func setAttributes() {
@@ -61,15 +62,56 @@ private extension TagsScrollView {
             make.edges.height.equalToSuperview()
         }
     }
+}
 
-    func setDelegate() {
-        delegate = self
+// MARK: - TagView
+
+private final class TagView: UIView {
+    private let label: UILabel = {
+        let label = UILabel()
+        label.text = "label"
+        label.textColor = .neutrals100
+        label.font = .systemFont(ofSize: 11.9)
+        return label
+    }()
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        configure()
+    }
+
+    @available(*, unavailable)
+    required init?(coder: NSCoder) {
+        fatalError()
+    }
+
+    func updateLabel(_ text: String) {
+        label.text = text
     }
 }
 
-extension TagsScrollView: UIScrollViewDelegate {
-    func scrollViewDidScroll(_ scrollView: UIScrollView) {
-        guard scrollView == self else { return }
-        scrollView.contentOffset.y = 0
+private extension TagView {
+    func configure() {
+        setAttributes()
+        setHierarchy()
+        setConstraints()
+    }
+
+    func setAttributes() {
+        backgroundColor = .neutrals600
+        layer.cornerRadius = 18
+        clipsToBounds = true
+    }
+
+    func setHierarchy() {
+        addSubview(label)
+    }
+
+    func setConstraints() {
+        label.snp.makeConstraints { make in
+            make.verticalEdges.equalToSuperview().inset(4)
+            make.directionalHorizontalEdges.equalToSuperview().inset(12)
+            make.height.equalTo(28)
+        }
     }
 }
